chore(deps): migrate rand family to the rand_core 0.10 line (rand 0.10 / rand_chacha 0.10 / rand_hc 0.5 / rand_distr 0.6 / getrandom 0.4) (#1193)
Moves the whole workspace off the rand-0.8/rand_core-0.6 generation in
lockstep, keeping deterministic key derivation bit-identical:

- Workspace deps: rand 0.10.2, rand_core 0.10.1, rand_chacha 0.10.0,
  rand_hc 0.5.1, rand_distr 0.6.0, getrandom 0.4.3.
- HC-128 verified unchanged between rand_hc 0.3.2 and 0.5.1 (identical
  core algorithm, identical eSTREAM test vectors incl. u64/byte-stream
  tests; BlockRng word/u64/byte semantics identical between rand_core
  0.6.4 and 0.10.1). All seeded goldens (x25519 fingerprint doctests,
  account/identity derivation, schnorrkel signing) pass unchanged; no
  fixture was regenerated.
- Trait migration: RngCore+CryptoRng bounds -> CryptoRng (which now
  subsumes Rng); CryptoRngCore -> CryptoRng; ext methods gen/gen_range/
  gen_bool -> random/random_range/random_bool via RngExt; thread_rng()
  -> rand::rng(); rand::distributions -> rand::distr; SliceRandom::choose
  -> IndexedRandom::choose; getrandom::getrandom() -> getrandom::fill().
- rand_core 0.10 dropped OsRng: added bth_util_from_random::OsRng, a
  zero-sized getrandom-backed infallible CryptoRng with the old
  panic-on-failure semantics, and pointed all former rand_core::OsRng /
  rand::rngs::OsRng users at it.
- Bridge shims: the dalek-5-era RngCompat (0.6 -> 0.10) in
  crypto/keys/x25519.rs is now redundant and removed (x25519-dalek 3
  takes workspace rngs directly). A reverse shim Rng06Compat
  (crypto/keys/src/compat.rs, workspace alias rand-core-v06) bridges
  workspace rngs into the still-rand_core-0.6 externals: schnorrkel-og
  transcript signing and bulletproofs-og range proofs. mc-rand keeps
  using its own 0.6 re-export; crypto/pq keeps its private rand_core 0.6
  pin for the ml-kem boundary (untouched).
- wasm: getrandom 0.2 "js" feature -> 0.4 "wasm_js" (cfg already set in
  .cargo/config.toml); wasm-signer keeps renamed 0.2/0.3 deps for the
  crypto/pq transitive generations. wasm32 target check passes.
- fuzz workspace bumped to the same line.
- .github/dependabot.yml: removed the rand-family ignore block this
  migration was gating.

Closes #1154

+++ botho-wallet/src/commands/init.rs
@@ -95,7 +95,7 @@ fn generate_mnemonic() -> Result<String> {
     }
 
     // Verify by asking for a random word
-    let verify_index = rand::random::<usize>() % 24;
+    let verify_index = rand::random::<u32>() as usize % 24;
     println!();
     print!("Verify: Enter word #{}: ", verify_index + 1);
     io::stdout().flush()?;
